@@ -1,22 +1,47 @@
 module github.com/Dreamacro/clash
 
-go 1.16
+go 1.21
 
 require (
-	github.com/Dreamacro/go-shadowsocks2 v0.1.7
-	github.com/go-chi/chi/v5 v5.0.3
-	github.com/go-chi/cors v1.2.0
-	github.com/go-chi/render v1.0.1
-	github.com/gofrs/uuid v4.0.0+incompatible
-	github.com/gorilla/websocket v1.4.2
-	github.com/miekg/dns v1.1.42
-	github.com/oschwald/geoip2-golang v1.5.0
-	github.com/sirupsen/logrus v1.8.1
-	github.com/stretchr/testify v1.7.0
-	go.uber.org/atomic v1.7.0
-	golang.org/x/crypto v0.0.0-20210506145944-38f3c27a63bf
-	golang.org/x/net v0.0.0-20210508051633-16afe75a6701
-	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
-	golang.org/x/sys v0.0.0-20210507161434-a76c4d0a0096
-	gopkg.in/yaml.v2 v2.4.0
+	github.com/Dreamacro/protobytes v0.0.0-20230617041236-6500a9f4f158
+	github.com/dlclark/regexp2 v1.10.0
+	github.com/go-chi/chi/v5 v5.0.10
+	github.com/go-chi/cors v1.2.1
+	github.com/go-chi/render v1.0.3
+	github.com/gofrs/uuid/v5 v5.0.0
+	github.com/gorilla/websocket v1.5.0
+	github.com/insomniacslk/dhcp v0.0.0-20230816195147-b3ca2534940d
+	github.com/mdlayher/netlink v1.7.2
+	github.com/miekg/dns v1.1.55
+	github.com/oschwald/geoip2-golang v1.9.0
+	github.com/samber/lo v1.38.1
+	github.com/sirupsen/logrus v1.9.3
+	github.com/stretchr/testify v1.8.4
+	github.com/vishvananda/netlink v1.2.1-beta.2.0.20230420174744-55c8b9515a01
+	go.etcd.io/bbolt v1.3.7
+	go.uber.org/atomic v1.11.0
+	go.uber.org/automaxprocs v1.5.3
+	golang.org/x/crypto v0.12.0
+	golang.org/x/net v0.14.0
+	golang.org/x/sync v0.3.0
+	golang.org/x/sys v0.11.0
+	gopkg.in/yaml.v3 v3.0.1
+)
+
+require (
+	github.com/ajg/form v1.5.1 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/google/go-cmp v0.5.9 // indirect
+	github.com/josharian/native v1.1.0 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/mdlayher/socket v0.4.1 // indirect
+	github.com/oschwald/maxminddb-golang v1.11.0 // indirect
+	github.com/pierrec/lz4/v4 v4.1.14 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/u-root/uio v0.0.0-20230220225925-ffce2a382923 // indirect
+	github.com/vishvananda/netns v0.0.0-20200728191858-db3c7e526aae // indirect
+	golang.org/x/exp v0.0.0-20220303212507-bbda1eaf7a17 // indirect
+	golang.org/x/mod v0.8.0 // indirect
+	golang.org/x/text v0.12.0 // indirect
+	golang.org/x/tools v0.6.0 // indirect
 )
