@@ -5,16 +5,14 @@ import (
 	"errors"
 )
 
-var (
-	// LogLevelMapping is a mapping for LogLevel enum
-	LogLevelMapping = map[string]LogLevel{
-		ERROR.String():   ERROR,
-		WARNING.String(): WARNING,
-		INFO.String():    INFO,
-		DEBUG.String():   DEBUG,
-		SILENT.String():  SILENT,
-	}
-)
+// LogLevelMapping is a mapping for LogLevel enum
+var LogLevelMapping = map[string]LogLevel{
+	ERROR.String():   ERROR,
+	WARNING.String(): WARNING,
+	INFO.String():    INFO,
+	DEBUG.String():   DEBUG,
+	SILENT.String():  SILENT,
+}
 
 const (
 	DEBUG LogLevel = iota
@@ -27,7 +25,7 @@ const (
 type LogLevel int
 
 // UnmarshalYAML unserialize LogLevel with yaml
-func (l *LogLevel) UnmarshalYAML(unmarshal func(interface{}) error) error {
+func (l *LogLevel) UnmarshalYAML(unmarshal func(any) error) error {
 	var tp string
 	unmarshal(&tp)
 	level, exist := LogLevelMapping[tp]
@@ -56,7 +54,7 @@ func (l LogLevel) MarshalJSON() ([]byte, error) {
 }
 
 // MarshalYAML serialize LogLevel with yaml
-func (l LogLevel) MarshalYAML() (interface{}, error) {
+func (l LogLevel) MarshalYAML() (any, error) {
 	return l.String(), nil
 }
 
