@@ -3,7 +3,7 @@ package constant
 import (
 	"net"
 
-	"github.com/gofrs/uuid"
+	"github.com/gofrs/uuid/v5"
 )
 
 type PlainContext interface {
