@@ -4,26 +4,34 @@ import (
 	"os"
 	P "path"
 	"path/filepath"
+	"strings"
 )
 
 const Name = "clash"
 
 // Path is used to get the configuration path
-var Path *path
-
-type path struct {
-	homeDir    string
-	configFile string
-}
-
-func init() {
+//
+// on Unix systems, `$HOME/.config/clash`.
+// on Windows, `%USERPROFILE%/.config/clash`.
+var Path = func() *path {
 	homeDir, err := os.UserHomeDir()
 	if err != nil {
 		homeDir, _ = os.Getwd()
 	}
 
 	homeDir = P.Join(homeDir, ".config", Name)
-	Path = &path{homeDir: homeDir, configFile: "config.yaml"}
+
+	if _, err = os.Stat(homeDir); err != nil {
+		if configHome, ok := os.LookupEnv("XDG_CONFIG_HOME"); ok {
+			homeDir = P.Join(configHome, Name)
+		}
+	}
+	return &path{homeDir: homeDir, configFile: "config.yaml"}
+}()
+
+type path struct {
+	homeDir    string
+	configFile string
 }
 
 // SetHomeDir is used to set the configuration path
@@ -53,10 +61,26 @@ func (p *path) Resolve(path string) string {
 	return path
 }
 
+// IsSubPath return true if path is a subpath of homedir
+func (p *path) IsSubPath(path string) bool {
+	homedir := p.HomeDir()
+	path = p.Resolve(path)
+	rel, err := filepath.Rel(homedir, path)
+	if err != nil {
+		return false
+	}
+
+	return !strings.Contains(rel, "..")
+}
+
 func (p *path) MMDB() string {
 	return P.Join(p.homeDir, "Country.mmdb")
 }
 
-func (p *path) Cache() string {
+func (p *path) OldCache() string {
 	return P.Join(p.homeDir, ".cache")
+}
+
+func (p *path) Cache() string {
+	return P.Join(p.homeDir, "cache.db")
 }
