@@ -6,13 +6,15 @@ import (
 
 	"github.com/Dreamacro/clash/component/resolver"
 	"github.com/Dreamacro/clash/config"
+	C "github.com/Dreamacro/clash/constant"
 	"github.com/Dreamacro/clash/hub/executor"
+	"github.com/Dreamacro/clash/listener"
 	"github.com/Dreamacro/clash/log"
-	P "github.com/Dreamacro/clash/proxy"
 	"github.com/Dreamacro/clash/tunnel"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/render"
+	"github.com/samber/lo"
 )
 
 func configRouter() http.Handler {
@@ -23,54 +25,29 @@ func configRouter() http.Handler {
 	return r
 }
 
-type configSchema struct {
-	Port        *int               `json:"port"`
-	SocksPort   *int               `json:"socks-port"`
-	RedirPort   *int               `json:"redir-port"`
-	TProxyPort  *int               `json:"tproxy-port"`
-	MixedPort   *int               `json:"mixed-port"`
-	AllowLan    *bool              `json:"allow-lan"`
-	BindAddress *string            `json:"bind-address"`
-	Mode        *tunnel.TunnelMode `json:"mode"`
-	LogLevel    *log.LogLevel      `json:"log-level"`
-	IPv6        *bool              `json:"ipv6"`
-}
-
 func getConfigs(w http.ResponseWriter, r *http.Request) {
 	general := executor.GetGeneral()
 	render.JSON(w, r, general)
 }
 
-func pointerOrDefault(p *int, def int) int {
-	if p != nil {
-		return *p
-	}
-
-	return def
-}
-
 func patchConfigs(w http.ResponseWriter, r *http.Request) {
-	general := &configSchema{}
-	if err := render.DecodeJSON(r.Body, general); err != nil {
+	general := struct {
+		Port        *int               `json:"port"`
+		SocksPort   *int               `json:"socks-port"`
+		RedirPort   *int               `json:"redir-port"`
+		TProxyPort  *int               `json:"tproxy-port"`
+		MixedPort   *int               `json:"mixed-port"`
+		AllowLan    *bool              `json:"allow-lan"`
+		BindAddress *string            `json:"bind-address"`
+		Mode        *tunnel.TunnelMode `json:"mode"`
+		LogLevel    *log.LogLevel      `json:"log-level"`
+		IPv6        *bool              `json:"ipv6"`
+	}{}
+	if err := render.DecodeJSON(r.Body, &general); err != nil {
 		render.Status(r, http.StatusBadRequest)
 		render.JSON(w, r, ErrBadRequest)
 		return
 	}
-
-	if general.AllowLan != nil {
-		P.SetAllowLan(*general.AllowLan)
-	}
-
-	if general.BindAddress != nil {
-		P.SetBindAddress(*general.BindAddress)
-	}
-
-	ports := P.GetPorts()
-	P.ReCreateHTTP(pointerOrDefault(general.Port, ports.Port))
-	P.ReCreateSocks(pointerOrDefault(general.SocksPort, ports.SocksPort))
-	P.ReCreateRedir(pointerOrDefault(general.RedirPort, ports.RedirPort))
-	P.ReCreateTProxy(pointerOrDefault(general.TProxyPort, ports.TProxyPort))
-	P.ReCreateMixed(pointerOrDefault(general.MixedPort, ports.MixedPort))
 
 	if general.Mode != nil {
 		tunnel.SetMode(*general.Mode)
@@ -84,16 +61,31 @@ func patchConfigs(w http.ResponseWriter, r *http.Request) {
 		resolver.DisableIPv6 = !*general.IPv6
 	}
 
+	if general.AllowLan != nil {
+		listener.SetAllowLan(*general.AllowLan)
+	}
+
+	if general.BindAddress != nil {
+		listener.SetBindAddress(*general.BindAddress)
+	}
+
+	ports := listener.GetPorts()
+	ports.Port = lo.FromPtrOr(general.Port, ports.Port)
+	ports.SocksPort = lo.FromPtrOr(general.SocksPort, ports.SocksPort)
+	ports.RedirPort = lo.FromPtrOr(general.RedirPort, ports.RedirPort)
+	ports.TProxyPort = lo.FromPtrOr(general.TProxyPort, ports.TProxyPort)
+	ports.MixedPort = lo.FromPtrOr(general.MixedPort, ports.MixedPort)
+
+	listener.ReCreatePortsListeners(*ports, tunnel.TCPIn(), tunnel.UDPIn())
+
 	render.NoContent(w, r)
 }
 
-type updateConfigRequest struct {
-	Path    string `json:"path"`
-	Payload string `json:"payload"`
-}
-
 func updateConfigs(w http.ResponseWriter, r *http.Request) {
-	req := updateConfigRequest{}
+	req := struct {
+		Path    string `json:"path"`
+		Payload string `json:"payload"`
+	}{}
 	if err := render.DecodeJSON(r.Body, &req); err != nil {
 		render.Status(r, http.StatusBadRequest)
 		render.JSON(w, r, ErrBadRequest)
@@ -112,6 +104,9 @@ func updateConfigs(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 	} else {
+		if req.Path == "" {
+			req.Path = C.Path.Config()
+		}
 		if !filepath.IsAbs(req.Path) {
 			render.Status(r, http.StatusBadRequest)
 			render.JSON(w, r, newError("path is not a absolute path"))
