@@ -8,16 +8,18 @@ import (
 	"path/filepath"
 	"runtime"
 	"syscall"
+	_ "time/tzdata"
 
 	"github.com/Dreamacro/clash/config"
 	C "github.com/Dreamacro/clash/constant"
 	"github.com/Dreamacro/clash/hub"
 	"github.com/Dreamacro/clash/hub/executor"
 	"github.com/Dreamacro/clash/log"
+
+	"go.uber.org/automaxprocs/maxprocs"
 )
 
 var (
-	flagset            map[string]bool
 	version            bool
 	testConfig         bool
 	homeDir            string
@@ -28,22 +30,18 @@ var (
 )
 
 func init() {
-	flag.StringVar(&homeDir, "d", "", "set configuration directory")
-	flag.StringVar(&configFile, "f", "", "specify configuration file")
-	flag.StringVar(&externalUI, "ext-ui", "", "override external ui directory")
-	flag.StringVar(&externalController, "ext-ctl", "", "override external controller address")
-	flag.StringVar(&secret, "secret", "", "override secret for RESTful API")
+	flag.StringVar(&homeDir, "d", os.Getenv("CLASH_HOME_DIR"), "set configuration directory")
+	flag.StringVar(&configFile, "f", os.Getenv("CLASH_CONFIG_FILE"), "specify configuration file")
+	flag.StringVar(&externalUI, "ext-ui", os.Getenv("CLASH_OVERRIDE_EXTERNAL_UI_DIR"), "override external ui directory")
+	flag.StringVar(&externalController, "ext-ctl", os.Getenv("CLASH_OVERRIDE_EXTERNAL_CONTROLLER"), "override external controller address")
+	flag.StringVar(&secret, "secret", os.Getenv("CLASH_OVERRIDE_SECRET"), "override secret for RESTful API")
 	flag.BoolVar(&version, "v", false, "show current version of clash")
 	flag.BoolVar(&testConfig, "t", false, "test configuration and exit")
 	flag.Parse()
-
-	flagset = map[string]bool{}
-	flag.Visit(func(f *flag.Flag) {
-		flagset[f.Name] = true
-	})
 }
 
 func main() {
+	maxprocs.Set(maxprocs.Logger(func(string, ...any) {}))
 	if version {
 		fmt.Printf("Clash %s %s %s with %s %s\n", C.Version, runtime.GOOS, runtime.GOARCH, runtime.Version(), C.BuildTime)
 		return
@@ -83,13 +81,13 @@ func main() {
 	}
 
 	var options []hub.Option
-	if flagset["ext-ui"] {
+	if externalUI != "" {
 		options = append(options, hub.WithExternalUI(externalUI))
 	}
-	if flagset["ext-ctl"] {
+	if externalController != "" {
 		options = append(options, hub.WithExternalController(externalController))
 	}
-	if flagset["secret"] {
+	if secret != "" {
 		options = append(options, hub.WithSecret(secret))
 	}
 
@@ -97,7 +95,20 @@ func main() {
 		log.Fatalln("Parse config error: %s", err.Error())
 	}
 
-	sigCh := make(chan os.Signal, 1)
-	signal.Notify(sigCh, syscall.SIGINT, syscall.SIGTERM)
-	<-sigCh
+	termSign := make(chan os.Signal, 1)
+	hupSign := make(chan os.Signal, 1)
+	signal.Notify(termSign, syscall.SIGINT, syscall.SIGTERM)
+	signal.Notify(hupSign, syscall.SIGHUP)
+	for {
+		select {
+		case <-termSign:
+			return
+		case <-hupSign:
+			if cfg, err := executor.ParseWithPath(C.Path.Config()); err == nil {
+				executor.ApplyConfig(cfg, true)
+			} else {
+				log.Errorln("Parse config error: %s", err.Error())
+			}
+		}
+	}
 }
