@@ -6,7 +6,7 @@ import (
 
 	C "github.com/Dreamacro/clash/constant"
 
-	"github.com/gofrs/uuid"
+	"github.com/gofrs/uuid/v5"
 	"go.uber.org/atomic"
 )
 
